+++ src/serve_session.rs
@@ -1,6 +1,6 @@
 use std::{
     collections::HashSet,
-    path::Path,
+    path::{Path, PathBuf},
     sync::{Arc, Mutex, MutexGuard},
     time::Instant,
 };
@@ -8,17 +8,18 @@ use std::{
 use crossbeam_channel::Sender;
 use memofs::Vfs;
 use rbx_dom_weak::RbxInstanceProperties;
+use thiserror::Error;
 
 use crate::{
     change_processor::ChangeProcessor,
     message_queue::MessageQueue,
-    project::Project,
+    project::{Project, ProjectError},
     session_id::SessionId,
     snapshot::{
         apply_patch_set, compute_patch_set, AppliedPatchSet, InstanceContext,
         InstancePropertiesWithMeta, PatchSet, PathIgnoreRule, RojoTree,
     },
-    snapshot_middleware::snapshot_from_vfs,
+    snapshot_middleware::{snapshot_from_vfs, SnapshotError},
 };
 
 /// Contains all of the state for a Rojo serve session.
@@ -47,15 +48,12 @@ pub struct ServeSession {
     /// diagnostics.
     start_time: Instant,
 
-    /// The root project for the serve session, if there was one defined.
+    /// The root project for the serve session.
     ///
     /// This will be defined if a folder with a `default.project.json` file was
     /// used for starting the serve session, or if the user specified a full
     /// path to a `.project.json` file.
-    ///
-    /// If `root_project` is None, values from the project should be treated as
-    /// their defaults.
-    root_project: Option<Project>,
+    root_project: Project,
 
     /// A randomly generated ID for this serve session. It's used to ensure that
     /// a client doesn't begin connecting to a different server part way through
@@ -86,9 +84,6 @@ pub struct ServeSession {
     tree_mutation_sender: Sender<PatchSet>,
 }
 
-/// Methods that need thread-safety bounds on VfsFetcher are limited to this
-/// block to prevent needing to spread Send + Sync + 'static into everything
-/// that handles ServeSession.
 impl ServeSession {
     /// Start a new serve session from the given in-memory filesystem and start
     /// path.
@@ -96,14 +91,17 @@ impl ServeSession {
     /// The project file is expected to be loaded out-of-band since it's
     /// currently loaded from the filesystem directly instead of through the
     /// in-memory filesystem layer.
-    pub fn new<P: AsRef<Path>>(vfs: Vfs, start_path: P) -> Self {
+    pub fn new<P: AsRef<Path>>(vfs: Vfs, start_path: P) -> Result<Self, ServeSessionError> {
         let start_path = start_path.as_ref();
         let start_time = Instant::now();
 
         log::trace!("Starting new ServeSession at path {}", start_path.display());
 
-        log::trace!("Loading project file from {}", start_path.display());
-        let root_project = Project::load_fuzzy(start_path).expect("TODO: Project load failed");
+        log::debug!("Loading project file from {}", start_path.display());
+        let root_project =
+            Project::load_fuzzy(start_path)?.ok_or_else(|| ServeSessionError::NoProjectFound {
+                path: start_path.to_owned(),
+            })?;
 
         let mut tree = RojoTree::new(InstancePropertiesWithMeta {
             properties: RbxInstanceProperties {
@@ -118,18 +116,18 @@ impl ServeSession {
 
         let mut instance_context = InstanceContext::default();
 
-        if let Some(project) = &root_project {
-            let rules = project.glob_ignore_paths.iter().map(|glob| PathIgnoreRule {
+        let rules = root_project
+            .glob_ignore_paths
+            .iter()
+            .map(|glob| PathIgnoreRule {
                 glob: glob.clone(),
-                base_path: project.folder_location().to_path_buf(),
+                base_path: root_project.folder_location().to_path_buf(),
             });
 
-            instance_context.add_path_ignore_rules(rules);
-        }
+        instance_context.add_path_ignore_rules(rules);
 
         log::trace!("Generating snapshot of instances from VFS");
-        let snapshot = snapshot_from_vfs(&instance_context, &vfs, &start_path)
-            .expect("snapshot failed")
+        let snapshot = snapshot_from_vfs(&instance_context, &vfs, &start_path)?
             .expect("snapshot did not return an instance");
 
         log::trace!("Computing initial patch set");
@@ -155,7 +153,7 @@ impl ServeSession {
             tree_mutation_receiver,
         );
 
-        Self {
+        Ok(Self {
             change_processor,
             start_time,
             session_id,
@@ -164,11 +162,9 @@ impl ServeSession {
             message_queue,
             tree_mutation_sender,
             vfs,
-        }
+        })
     }
-}
 
-impl ServeSession {
     pub fn tree_handle(&self) -> Arc<Mutex<RojoTree>> {
         Arc::clone(&self.tree)
     }
@@ -181,6 +177,7 @@ impl ServeSession {
         self.tree_mutation_sender.clone()
     }
 
+    #[allow(unused)]
     pub fn vfs(&self) -> &Vfs {
         &self.vfs
     }
@@ -193,16 +190,12 @@ impl ServeSession {
         self.session_id
     }
 
-    pub fn project_name(&self) -> Option<&str> {
-        self.root_project
-            .as_ref()
-            .map(|project| project.name.as_str())
+    pub fn project_name(&self) -> &str {
+        &self.root_project.name
     }
 
     pub fn project_port(&self) -> Option<u16> {
-        self.root_project
-            .as_ref()
-            .and_then(|project| project.serve_port)
+        self.root_project.serve_port
     }
 
     pub fn start_time(&self) -> Instant {
@@ -210,217 +203,28 @@ impl ServeSession {
     }
 
     pub fn serve_place_ids(&self) -> Option<&HashSet<u64>> {
-        self.root_project
-            .as_ref()
-            .and_then(|project| project.serve_place_ids.as_ref())
+        self.root_project.serve_place_ids.as_ref()
     }
 }
 
-/// This module is named to trick Insta into naming the resulting snapshots
-/// correctly.
-///
-/// See https://github.com/mitsuhiko/insta/issues/78
-#[cfg(test)]
-mod serve_session {
-    use super::*;
+#[derive(Debug, Error)]
+pub enum ServeSessionError {
+    #[error(
+        "Rojo requires a project file, but no project file was found in path {}\n\
+        See https://rojo.space/docs/ for guides and documentation.",
+        .path.display()
+    )]
+    NoProjectFound { path: PathBuf },
 
-    use std::{path::PathBuf, time::Duration};
+    #[error(transparent)]
+    Project {
+        #[from]
+        source: ProjectError,
+    },
 
-    use maplit::hashmap;
-    use memofs::{InMemoryFs, VfsEvent, VfsSnapshot};
-    use rojo_insta_ext::RedactionMap;
-    use tokio::{runtime::Runtime, timer::Timeout};
-
-    use crate::tree_view::view_tree;
-
-    #[test]
-    fn just_folder() {
-        let mut imfs = InMemoryFs::new();
-        imfs.load_snapshot("/foo", VfsSnapshot::empty_dir())
-            .unwrap();
-
-        let vfs = Vfs::new(imfs);
-
-        let session = ServeSession::new(vfs, "/foo");
-
-        let mut rm = RedactionMap::new();
-        insta::assert_yaml_snapshot!(view_tree(&session.tree(), &mut rm));
-    }
-
-    #[test]
-    fn project_with_folder() {
-        let mut imfs = InMemoryFs::new();
-        imfs.load_snapshot(
-            "/foo",
-            VfsSnapshot::dir(hashmap! {
-                "default.project.json" => VfsSnapshot::file(r#"
-                    {
-                        "name": "HelloWorld",
-                        "tree": {
-                            "$path": "src"
-                        }
-                    }
-                "#),
-                "src" => VfsSnapshot::dir(hashmap! {
-                    "hello.txt" => VfsSnapshot::file("Hello, world!"),
-                }),
-            }),
-        )
-        .unwrap();
-
-        let vfs = Vfs::new(imfs);
-
-        let session = ServeSession::new(vfs, "/foo");
-
-        let mut rm = RedactionMap::new();
-        insta::assert_yaml_snapshot!(view_tree(&session.tree(), &mut rm));
-    }
-
-    #[test]
-    fn script_with_meta() {
-        let mut imfs = InMemoryFs::new();
-        imfs.load_snapshot(
-            "/root",
-            VfsSnapshot::dir(hashmap! {
-                "test.lua" => VfsSnapshot::file("This is a test."),
-                "test.meta.json" => VfsSnapshot::file(r#"{ "ignoreUnknownInstances": true }"#),
-            }),
-        )
-        .unwrap();
-
-        let vfs = Vfs::new(imfs);
-
-        let session = ServeSession::new(vfs, "/root");
-
-        let mut rm = RedactionMap::new();
-        insta::assert_yaml_snapshot!(view_tree(&session.tree(), &mut rm));
-    }
-
-    #[test]
-    fn change_txt_file() {
-        let mut imfs = InMemoryFs::new();
-        imfs.load_snapshot("/foo.txt", VfsSnapshot::file("Hello!"))
-            .unwrap();
-
-        let vfs = Vfs::new(imfs.clone());
-
-        let session = ServeSession::new(vfs, "/foo.txt");
-
-        let mut rm = RedactionMap::new();
-        insta::assert_yaml_snapshot!(
-            "change_txt_file_before",
-            view_tree(&session.tree(), &mut rm)
-        );
-
-        imfs.load_snapshot("/foo.txt", VfsSnapshot::file("World!"))
-            .unwrap();
-
-        let receiver = session.message_queue().subscribe_any();
-
-        imfs.raise_event(VfsEvent::Write(PathBuf::from("/foo.txt")));
-
-        let receiver = Timeout::new(receiver, Duration::from_millis(200));
-
-        let mut rt = Runtime::new().unwrap();
-        let result = rt.block_on(receiver).unwrap();
-
-        insta::assert_yaml_snapshot!("change_txt_file_patch", rm.redacted_yaml(result));
-        insta::assert_yaml_snapshot!("change_txt_file_after", view_tree(&session.tree(), &mut rm));
-    }
-
-    #[test]
-    fn change_script_meta() {
-        let mut imfs = InMemoryFs::new();
-        imfs.load_snapshot(
-            "/root",
-            VfsSnapshot::dir(hashmap! {
-                "test.lua" => VfsSnapshot::file("This is a test."),
-                "test.meta.json" => VfsSnapshot::file(r#"{ "ignoreUnknownInstances": true }"#),
-            }),
-        )
-        .unwrap();
-
-        let vfs = Vfs::new(imfs.clone());
-
-        let session = ServeSession::new(vfs, "/root");
-
-        let mut rm = RedactionMap::new();
-        insta::assert_yaml_snapshot!(
-            "change_script_meta_before",
-            view_tree(&session.tree(), &mut rm)
-        );
-
-        imfs.load_snapshot(
-            "/root/test.meta.json",
-            VfsSnapshot::file(r#"{ "ignoreUnknownInstances": false }"#),
-        )
-        .unwrap();
-
-        let receiver = session.message_queue().subscribe_any();
-
-        imfs.raise_event(VfsEvent::Write(PathBuf::from("/root/test.meta.json")));
-
-        let receiver = Timeout::new(receiver, Duration::from_millis(200));
-
-        let mut rt = Runtime::new().unwrap();
-        let result = rt.block_on(receiver).unwrap();
-
-        insta::assert_yaml_snapshot!("change_script_meta_patch", rm.redacted_yaml(result));
-        insta::assert_yaml_snapshot!(
-            "change_script_meta_after",
-            view_tree(&session.tree(), &mut rm)
-        );
-    }
-
-    #[test]
-    fn change_file_in_project() {
-        let mut imfs = InMemoryFs::new();
-        imfs.load_snapshot(
-            "/foo",
-            VfsSnapshot::dir(hashmap! {
-                "default.project.json" => VfsSnapshot::file(r#"
-                    {
-                        "name": "change_file_in_project",
-                        "tree": {
-                            "$className": "Folder",
-
-                            "Child": {
-                                "$path": "file.txt"
-                            }
-                        }
-                    }
-                "#),
-                "file.txt" => VfsSnapshot::file("initial content"),
-            }),
-        )
-        .unwrap();
-
-        let vfs = Vfs::new(imfs.clone());
-
-        let session = ServeSession::new(vfs, "/foo");
-
-        let mut rm = RedactionMap::new();
-        insta::assert_yaml_snapshot!(
-            "change_file_in_project_before",
-            view_tree(&session.tree(), &mut rm)
-        );
-
-        imfs.load_snapshot("/foo/file.txt", VfsSnapshot::file("Changed!"))
-            .unwrap();
-
-        let receiver = session.message_queue().subscribe_any();
-
-        imfs.raise_event(VfsEvent::Write(PathBuf::from("/foo/file.txt")));
-
-        let receiver = Timeout::new(receiver, Duration::from_millis(200));
-
-        let mut rt = Runtime::new().unwrap();
-        let result = rt.block_on(receiver).unwrap();
-
-        insta::assert_yaml_snapshot!("change_file_in_project_patch", rm.redacted_yaml(result));
-        insta::assert_yaml_snapshot!(
-            "change_file_in_project_after",
-            view_tree(&session.tree(), &mut rm)
-        );
-    }
+    #[error(transparent)]
+    Snapshot {
+        #[from]
+        source: SnapshotError,
+    },
 }