+++ src/snapshot_middleware/mod.rs
@@ -6,6 +6,7 @@
 mod csv;
 mod dir;
 mod error;
+mod json;
 mod json_model;
 mod lua;
 mod meta_file;
@@ -17,20 +18,27 @@ mod rbxmx;
 mod txt;
 mod util;
 
-pub use self::error::*;
-
 use std::path::Path;
 
 use memofs::Vfs;
 
-use self::middleware::{SnapshotInstanceResult, SnapshotMiddleware};
-use self::{
-    csv::SnapshotCsv, dir::SnapshotDir, json_model::SnapshotJsonModel, lua::SnapshotLua,
-    project::SnapshotProject, rbxlx::SnapshotRbxlx, rbxm::SnapshotRbxm, rbxmx::SnapshotRbxmx,
-    txt::SnapshotTxt,
-};
 use crate::snapshot::InstanceContext;
 
+use self::{
+    csv::SnapshotCsv,
+    dir::SnapshotDir,
+    json::SnapshotJson,
+    json_model::SnapshotJsonModel,
+    lua::SnapshotLua,
+    middleware::{SnapshotInstanceResult, SnapshotMiddleware},
+    project::SnapshotProject,
+    rbxlx::SnapshotRbxlx,
+    rbxm::SnapshotRbxm,
+    rbxmx::SnapshotRbxmx,
+    txt::SnapshotTxt,
+};
+
+pub use self::error::*;
 pub use self::project::snapshot_project_node;
 
 macro_rules! middlewares {
@@ -65,5 +73,6 @@ middlewares! {
     SnapshotLua,
     SnapshotCsv,
     SnapshotTxt,
+    SnapshotJson,
     SnapshotDir,
 }